fix: always replace common.h include in .c file in non-dry-run

+++ tools/restructure.py
@@ -170,17 +170,30 @@ def main():
 
 #endif // __{define_name}_H__
 """)
-                # Open current file, replace common.h include with overlays header (newly created file) import
-                file_content = ""
-                with open(target_file_name, "r") as f:
-                    file_content = f.read()
-                if file_content == "":
-                    print(f"Failed to read {target_file_name}, could not replace #include statement")
-                    continue
+        if DRY_RUN is False:
+            # Open current file
+            file_content = ""
+            with open(target_file_name, "r") as f:
+                file_content = f.read()
+            if file_content == "":
+                print(f"Failed to read {target_file_name}, could not replace #include statement")
+                continue
 
-                file_content = file_content.replace('#include "common.h"', f'#include "{expected_h_file_name.replace(PROJECT_ROOT + "/src/", "")}"')
-                with open(target_file_name, "w") as f:
-                    f.write(file_content)
+            if "common.h" in file_content:
+                # Replace common.h include with overlays header (newly created file) import
+                header_include_path = expected_h_file_name.replace(PROJECT_ROOT + "/src/", "")
+                print(f"Replacing common.h with {header_include_path}")
+                file_content = file_content.replace('#include "common.h"', f'#include "{header_include_path}"')
+
+            # Replace GLOBAL_ASM paths to match new structure
+            current_asm_path = file.replace(PROJECT_ROOT + "/src/", "asm/nonmatchings/").replace(".c", "/")
+            target_asm_path = target_file_name.replace(PROJECT_ROOT + "/src/", "asm/nonmatchings/").replace(".c", "/")
+            print(f"Replacing {current_asm_path} with {target_asm_path}")
+            file_content = file_content.replace(current_asm_path, target_asm_path)
+
+            # Write back file content
+            with open(target_file_name, "w") as f:
+                f.write(file_content)
 
         if DRY_RUN is False:
             current_asm_path = file.replace(PROJECT_ROOT + "/src/", "asm/nonmatchings/").replace(".c", "/")
